Increase motor speed
Modifies the Sprint 2 file to use the AccelStepper library to increase the speed of the motors as they dispense drinks. Currently set to dispense drinks 1 and 2 for 30 seconds, and for drink 3 we are dispensing ingredient 2 for 45 seconds, and ingredient 1 for a minute.

TODO:
*Figure out how much water is being Dispensed per amount of time.
*Find flow rate of liquid at max speed
*Find flow rate of liquid during acceleration.
*Find flow rate of liquid during deceleration.

NOTE: At current settings, it takes ~17 seconds to hit max speed, and thus also begins to decelerate ~17 seconds before the end of the time. So drink dispensing duration must be of at least 34 seconds.

Also, the motor still run pretty hot.